Refactor/1839/refactor test for category item component1 (#1860)
* refactored test for category-item component

* refactored test

* re-refactored test

* retry CI-test

* resolved remark
